Set renderedInline on the AbstractReviewable model.
We have a `renderedInline` attribute on both the `AbstractReviewable` model
and `AbstractReviewableView` JS objects for our review UIs. This
attribute says whether the review UI is being rendered inline or not. We
were properly setting this on `AbstractReviewableView`, however we
weren't setting it on `AbstractReviewable`. This change fixes that.

We also add a `ReviewUI._inline` attribute that gets set along with
the `request` attribute. This way any method on the review UI can
access the inline information.

Testing Done:
- Confirmed that the attribute is set on the `AbstractReviewable`,
  viewed an inline review UI and a standalone one.
- Confirmed that the attribute is still properly set on
  `AbstractReviewableView`.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14627/