Fix required fields for the Assembla hosting service.
A while back we had made changes to the hosting services to require
keyword arguments for `check_repository`. In the Assembla
implementation, this change ended up changing the `assembla_project_id`
field to not include a default value. This field is required when using
a Perforce repository, but is not when using SVN.

In practice, the repository form will always provide this, but we had a
unit test that didn't. To prefer existing compatibility, I've given this
field a `None` default again, and raise a `ValueError` if it's not
provided in the Perforce case.

While in here I also noticed that the skip messages for unit tests with
SVN repositories were wrong.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14702/