Optimize HideActionHook and improve related typing.
`HideActionHook` now stores the IDs as a set instead of a list, which is
faster for lookup in the case where multiple IDs are provided.

This and `ActionHook` have also been updated to take and store immutable
types, which makes it easier to take in values coming from another
source that may be immutable.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14664/