Add registered action attachment points with rendering control.
When actions were first introduced, an action could specify an
attachment point that they wanted to render into. These were strings
that corresponded to a template tag call, rendering all actions attached
to that point at that location.

This change builds upon that concept by introducing a formal
registration of attachment points. These provide a level of control for
the rendering and state of any actions placed within them, allowing, for
example, one attachment point to render actions as a list of links and
another to render those same actions as a list of buttons.

Each attachment point has a default renderer that's used if the action
being rendered doesn't specify its own default. The intended approach
going forward is to have actions only specify a renderer if they have a
strong need for certain rendering behavior, but to otherwise leave that
up to the parent action group, attachment point, or to the (coming soon)
placement configuration.

Attachment points may also specify a list of actions to directly
include, in the order included. These may be actions registered to this
attachment point, or they may not be. This allows an attachment point to
easily include a set of actions rendered under the attachment point's
control for special extension use cases.

Testing Done:
Unit tests pass.

Tested all current actions in the UI. They all appear and behave as
expected.

Tested with in-progress changes that move Quick Access actions to this
new system, and verified they rendered as expected.

Reviewed at https://reviews.reviewboard.org/r/14667/