Move more template context to the action renderers.
In preparation for the final changes in separating actions from their
views, this change pulls out some of the view-specific state from the
actions and sets them in the renderers instead. This includes the
`has_parent` flag and `dom_element_id` (which previously required a call
to `get_dom_element_id()`).

`BaseAction.get_dom_element_id()` and the JavaScript `Action` model's
`domID` attribute are now both deprecated.

There's no good way to shim `domID`, as it's an action model attribute
and is fundamentally incompatible with per-view dom IDs. We only use
this ourselves for menus, and it doesn't appear we ever actually do
anything with it anywhere, so we're just scrapping it.

Testing Done:
Unit tests pass.

Tested that all actions displayed and worked correctly.

Reviewed at https://reviews.reviewboard.org/r/14671/