Organize children in grouped actions by attachment point ID.
Now that actions can live in multiple attachment points, and may or may
not have the same parent across them, we no longer want to represent
only a single list of children.

The `children` attribute for `GroupAction` is now a mapping of
attachment point IDs to lists of children. This is set during render of
the model. By grouping it this way, the action will factor in the
correct children for a given group.

This does not impact how actions are configured when defined. It's
simply the representation to the rendering and view code.

Testing Done:
Unit tests pass.

Tested that all menu actions contained the right children.

Reviewed at https://reviews.reviewboard.org/r/14672/