Allow action views to maintain visibility independently of the action.
When actions and action views had a one-to-one relationship, we used the
`visible` attribute to manage visibility of the action. This is no
longer the case, as we have actions that may have views in more than one
location but where one of those locations should not necessarily be
visible.

For example, the Ship It action should always be visible in the Review
menu, but the one in the Quick Access area may need to be hidden.

We now track the visibility state separately from the attribute. If the
attribute is `true`, then the view manages its own visibility. If it's
`false`, then the view will always be hidden.

Testing Done:
Unit tests passed.

Tested this with in-progress work on reusing existing actions for
Quick Access.

Reviewed at https://reviews.reviewboard.org/r/14673/