Update all existing actions to use the new placements.
This updates all the built-in actions in the UI to set placements
instead of attachments and parent IDs. This brings them up to the modern
interface, and enables the actions to be used in other attachment points
as needed.

Testing Done:
Unit tests pass.

Tested all these actions in the UI.

Reviewed at https://reviews.reviewboard.org/r/14715/