Add action renderers for sidebars.
This introduces two new renderers used for sidebars: One for groups
(which recurses into subgroups), and one for items.

The renderers fit into the existing `rb-c-sidebar` styling. This has
been updated to support subgroups in the sidebar, which appear with a
bold (but not uppercase) header and nested items.

This in preparation for action registration for admin sidebar items,
which will allow extensions to augment the sidebar.

Testing Done:
Unit tests pass.

Tested with an upcoming change converting the admin UI to the sidebar.

Reviewed at https://reviews.reviewboard.org/r/14716/