Fix setting NODE_PATH.
A last-minute bug snuck into the settings for NODE_PATH while I was
moving the jquery-ui stuff over to the build-media script, causing the
generator for node_modules paths to be evaluated twice. This resulted in
NODE_PATH being set to an empty string. This change removes the code
that was supposed to be removed.

Testing Done:
Built static media successfully.

Reviewed at https://reviews.reviewboard.org/r/14744/